Added nix support (WIP)
Added a nix flake and derivations for pyqir and pytket-qir,
using the tket flake from github for pytket module inclusion.
The tket flake's nixpkgs is overridden with the current package's
flake to avoid unnecessary download, and to help spot compatibility
problems more easily.

Currently mypy tests are disabled because there are several
type errors that appear. I do not know if these are expected,
or if it is a problem with this derivation. I have also tested
with my `feature/nix-support-add-mypy` branch of tket and,
although the errors are reduced, there are some outstanding
mypy issues coming from within this repository (at least, on
the nix build). Advice would be appreciated on this aspect.

Added result directory to .gitignore, as this can be created
by nix during development.